Rename InputProviderHelper to InputProviderUtil and move it to org.thunlp.util
Since this is the right place/name for a util class.
